Store LTML dimensions in float32

+++ ltml/corners_test.go
@@ -16,7 +16,7 @@ func TestCorners_SetAll_default(t *testing.T) {
 
 func TestCorners_SetAll_1(t *testing.T) {
 	var corners Corners
-	var expected = []float64{3.0}
+	var expected = []float32{3.0}
 	corners.SetAll("3", "")
 	if len(corners) != len(expected) {
 		t.Errorf("Expected %d, got %d", len(expected), len(corners))
@@ -30,7 +30,7 @@ func TestCorners_SetAll_1(t *testing.T) {
 
 func TestCorners_SetAll_2(t *testing.T) {
 	var corners Corners
-	var expected = []float64{4.0, 5.0}
+	var expected = []float32{4.0, 5.0}
 	corners.SetAll("4 5", "")
 	if len(corners) != len(expected) {
 		t.Errorf("Expected %d, got %d", len(expected), len(corners))
@@ -44,7 +44,7 @@ func TestCorners_SetAll_2(t *testing.T) {
 
 func TestCorners_SetAll_4(t *testing.T) {
 	var corners Corners
-	var expected = []float64{4.0, 5.0, 6.0, 7.0}
+	var expected = []float32{4.0, 5.0, 6.0, 7.0}
 	corners.SetAll("4 5 6 7", "")
 	if len(corners) != len(expected) {
 		t.Errorf("Expected %d, got %d", len(expected), len(corners))
@@ -58,7 +58,7 @@ func TestCorners_SetAll_4(t *testing.T) {
 
 func TestCorners_SetAll_8(t *testing.T) {
 	var corners Corners
-	var expected = []float64{6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0}
+	var expected = []float32{6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0}
 	corners.SetAll("6 7 8 9 10 11 12 13", "")
 	if len(corners) != len(expected) {
 		t.Errorf("Expected %d, got %d", len(expected), len(corners))
